requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
sqlalchemy==2.0.48
pymysql==1.1.2
python-dotenv==1.2.2